backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
APScheduler==3.10.4
httpx==0.25.1
python-dateutil==2.8.2
requests==2.31.0
# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Fix de compatibilidade: bcrypt 4 remove __about__; pin para 3.2.0
bcrypt==3.2.0
# Rate Limiting
slowapi==0.1.9
